feat: collection folder structure changed
