+++ packages/at_client/lib/at_client.dart
@@ -19,9 +19,9 @@ export 'package:at_client/src/key_stream/key_stream.dart';
 export 'package:at_client/src/service/sync/sync_conflict.dart';
 export 'package:at_commons/at_commons.dart';
 export 'package:at_client/src/listener/sync_progress_listener.dart';
-export 'package:at_client/src/at_collection/model/at_collection_impl.dart';
+export 'package:at_client/src/at_collection/at_collection_impl.dart';
 export 'package:at_client/src/at_collection/model/at_collection_model.dart';
-export 'package:at_client/src/at_collection/model/at_collection_spec.dart';
-export 'package:at_client/src/at_collection/model/at_collection_model_methods_mixin.dart';
+export 'package:at_client/src/at_collection/at_collection_spec.dart';
+export 'package:at_client/src/at_collection/at_collection_model_methods_mixin.dart';
 export 'package:at_client/src/at_collection/model/at_share_operation.dart';
 export 'package:at_client/src/at_collection/model/at_unshare_operation.dart';
